+++ backend/api/src/DB/news.go
@@ -1,0 +1,1 @@
+package db
